+++ src/tiled/session.cpp
@@ -222,6 +222,9 @@ static QString lastPathKey(Session::FileType fileType)
     case Session::WorldFile:
         key.append(QLatin1String("worldFilePath"));
         break;
+    case Session::ShortcutSettingsFile:
+        key.append(QLatin1String("shortcutSettingsFilePath"));
+        break;
     }
 
     return key;