+++ src/tiled/session.h
@@ -169,6 +169,7 @@ public:
         ObjectTypesFile,
         WorkingDirectory,
         WorldFile,
+        ShortcutSettingsFile,
     };
 
     QString lastPath(FileType fileType,