+++ NEWS.md
@@ -3,6 +3,7 @@
 * Keep references between objects when copy/pasting (#3361)
 * Improved default translation used in case of multiple options
 * Fixed 'Add Variation' action in Tile Stamps context menu (#3362)
+* Fixed importing of removed shortcuts (#3367)
 
 ### Tiled 1.8.5 (17 May 2022)